C file #6 - compliling errors depend of the gcc option

+++ 0x00-hello_world/6-size.c
@@ -8,7 +8,7 @@ int main(void)
 	int szchr = sizeof(char);
 	int szint = sizeof(int);
 	int szlnt = sizeof(long int);
-	int szlln = sizeof(long int);
+	int szlln = sizeof(long long);
 	int szflt = sizeof(float);
 
 	printf("Size of a char: %d byte(s)\n", szchr);
